Extract method to trait
